Move uqi into its own package
