+++ demo-app/src/lib/mysql-uqi-client.ts
@@ -6,7 +6,7 @@ import {
   type UqiMappedType,
   type UqiResult,
   type UqiScalar,
-} from '@open-truss/open-truss'
+} from '@open-truss/uqi'
 import mysql, { type Connection, type FieldPacket } from 'mysql2/promise'
 import { URL } from 'url'
