+++ packages/open-truss/src/index.ts
@@ -3,8 +3,6 @@ export * from './configuration'
 export * from './hooks'
 export * from './shims'
 export * from './signals'
-export * from './uqi/iterator'
-export * from './uqi/uqi'
 export * from './utils/describe-zod'
 export * from './utils/format'
 export * from './utils/yaml'
