+++ packages/uqi/src/iterator.ts
